Move tool window into UI namespace
